fix runtime_cyclesPerSecond rename
