+++ godeltaprof/compat/stub_go23_test.go
@@ -1,17 +1,23 @@
-//go:build go1.16 && !go1.21
-// +build go1.16,!go1.21
+//go:build go1.23
+// +build go1.23
 
 package compat
 
 import "testing"
 
 func TestRuntimeFrameSymbolName(t *testing.T) {
+	checkSignature(t, "runtime/pprof",
+		"runtime_FrameSymbolName",
+		"func runtime/pprof.runtime_FrameSymbolName(f *runtime.Frame) string")
 	checkSignature(t, "github.com/grafana/pyroscope-go/godeltaprof/internal/pprof",
 		"runtime_FrameSymbolName",
 		"func github.com/grafana/pyroscope-go/godeltaprof/internal/pprof.runtime_FrameSymbolName(f *runtime.Frame) string")
 }
 
 func TestRuntimeFrameStartLine(t *testing.T) {
+	checkSignature(t, "runtime/pprof",
+		"runtime_FrameStartLine",
+		"func runtime/pprof.runtime_FrameStartLine(f *runtime.Frame) int")
 	checkSignature(t, "github.com/grafana/pyroscope-go/godeltaprof/internal/pprof",
 		"runtime_FrameStartLine",
 		"func github.com/grafana/pyroscope-go/godeltaprof/internal/pprof.runtime_FrameStartLine(f *runtime.Frame) int")
@@ -27,9 +33,9 @@ func TestRuntimeExpandFinalInlineFrame(t *testing.T) {
 }
 
 func TestRuntimeCyclesPerSecond(t *testing.T) {
-	checkSignature(t, "runtime/pprof",
-		"runtime_cyclesPerSecond",
-		"func runtime/pprof.runtime_cyclesPerSecond() int64")
+	checkSignature(t, "runtime",
+		"pprof_cyclesPerSecond",
+		"func runtime.pprof_cyclesPerSecond() int64")
 	checkSignature(t, "github.com/grafana/pyroscope-go/godeltaprof/internal/pprof",
 		"runtime_cyclesPerSecond",
 		"func github.com/grafana/pyroscope-go/godeltaprof/internal/pprof.runtime_cyclesPerSecond() int64")
