+++ godeltaprof/internal/pprof/stub_go22.go
@@ -1,5 +1,5 @@
-//go:build go1.21
-// +build go1.21
+//go:build go1.21 && !go1.23
+// +build go1.21,!go1.23
 
 package pprof
 
@@ -19,3 +19,9 @@ func runtime_FrameStartLine(f *runtime.Frame) int
 //go:noescape
 //go:linkname runtime_FrameSymbolName runtime/pprof.runtime_FrameSymbolName
 func runtime_FrameSymbolName(f *runtime.Frame) string
+
+//go:linkname runtime_expandFinalInlineFrame runtime/pprof.runtime_expandFinalInlineFrame
+func runtime_expandFinalInlineFrame(stk []uintptr) []uintptr
+
+//go:linkname runtime_cyclesPerSecond runtime/pprof.runtime_cyclesPerSecond
+func runtime_cyclesPerSecond() int64
